core: move patches and extensions to namespace
